+++ tests/integration/main.rs
@@ -13,6 +13,7 @@ mod grep;
 mod helpers;
 mod import_doing;
 mod import_timing;
+mod interactive;
 mod last;
 mod mark;
 mod meanwhile;